A couple of minor fixes
* removed a useless 'explicit'
* trust to the implicit (double) conversion

+++ src/store_rebuild.cc
@@ -138,8 +138,8 @@ StoreRebuildFinalize()
         debugs(20, DBG_IMPORTANT, "  " << std::setw(7) << counts.cancelcount  << " Objects cancelled.");
         debugs(20, DBG_IMPORTANT, "  " << std::setw(7) << counts.dupcount  << " Duplicate URLs purged.");
         debugs(20, DBG_IMPORTANT, "  " << std::setw(7) << counts.clashcount  << " Swapfile clashes avoided.");
-        debugs(20, DBG_IMPORTANT, "  Took "<< std::setw(3)<< std::setprecision(2) << dt << " seconds ("<< std::setw(6) <<
-               ((double) counts.objcount / (dt > 0.0 ? dt : 1.0)) << " objects/sec).");
+        debugs(20, DBG_IMPORTANT, "  Took " << std::setw(3) << std::setprecision(2) << dt << " seconds (" << std::setw(6) <<
+               (counts.objcount / (dt > 0.0 ? dt : 1.0)) << " objects/sec).");
     }
 
     delete RebuildProgress;
